+++ projects/igniteui-angular/src/lib/grids/row.directive.ts
@@ -100,16 +100,31 @@ export class IgxRowDirective<T extends IgxGridBaseDirective & GridType> implemen
     @ViewChild(forwardRef(() => IgxCheckboxComponent), { read: IgxCheckboxComponent })
     public checkboxElement: IgxCheckboxComponent;
 
+    @ViewChildren('cell')
+    protected _cells: QueryList<any>;
+
     /**
-     * The rendered cells in the row component.
+     * Gets the rendered cells in the row component.
      *
      * ```typescript
      * // get the cells of the third selected row
      * let selectedRowCells = this.grid.selectedRows[2].cells;
      * ```
      */
-    @ViewChildren(forwardRef(() => IgxGridCellComponent))
-    public cells: QueryList<IgxGridCellComponent>;
+    public get cells() {
+        const res = new QueryList<any>();
+        if (!this._cells) {
+            return res;
+        }
+        const cList = this._cells.filter((item) => item.nativeElement.parentElement !== null)
+        .sort((item1, item2) => item1.column.visibleIndex - item2.column.visibleIndex);
+        res.reset(cList);
+        return res;
+    }
+
+    public set cells(cells) {
+
+    }
 
     /**
      * @hidden
@@ -361,6 +376,11 @@ export class IgxRowDirective<T extends IgxGridBaseDirective & GridType> implemen
      */
     public delete() {
         this.grid.deleteRowById(this.rowID);
+    }
+
+    public isCellActive(visibleColumnIndex) {
+        const node = this.grid.navigation.activeNode;
+        return node ? node.row === this.index && node.column === visibleColumnIndex : false;
     }
 
     /**